Fix broken pressure/temperature handling
In the original version, I had gotten lazy and just pulled the temperature and
pressure out of the drag table as soon as it was specified.  This imposed
a stupid condition on the user to specify the drag function earlier on the
command line than any temperature or pressure setting.

This fix makes us defer pulling temperature and pressure out of the drag table
until after the command line is parsed out.  Only if no temperature or pressure
is specified do temperature and pressure get set from the drag function table.