Silence warnings from LaTeX
The "ballistics.tex" file was originally intended to be a summary of
what I'd learned about exterior ballistics in order to write this
program.  I never got past the first section, as I gave up on the idea
of sharing this with anyone.

Now that free, high-quality ballistics programs can be found simply by
flinging a brick and hitting one, there's even less of a point to
sharing this than ever.  But here we are.

And I found that when I tried to build this document pdflatex screamed
at me that I was using "documentstyle" instead of "documentclass".
Fortunately, I wasn't doing anything fancy with the style and simply
changing the former into the latter is good enough to shut LaTeX up.