Fix broken computation of drop at target distance
The previous commit output the ys[i] drop at the target distance, but
this has already been adjusted to take into account the behavior at
the zero distance.  As such, it was not the drop from the bore line
and not useful for the comparisons for which it was implemented.

This one is.